Add almacen_producto, servicios, seed db

+++ database/migrations/2025_03_03_163050_create_productos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nombre', 255);
             $table->decimal('precio_compra', 10, 2);
             $table->decimal('precio_venta', 10, 2);
-            $table->decimal('iva', 2, 2);
+            $table->unsignedInteger('iva');
             $table->string('descripcion', 255)->nullable();
             $table->timestamps();
         });
